pyproject.toml
[project]
name = "redirect"
description = "Redirect service"
version = "1.0.0"
readme = "README.md"
dependencies = [
    "aiodns==3.1.1",
    "aiohttp==3.9.1",
    "Brotli==1.1.0",
    "pydantic==2.5.3",
    "pydantic-core==2.14.6",
    "pydantic-settings==2.1.0",
    "roundrobin==0.0.4",
]
requires-python = ">=3.12"

[project.optional-dependencies]
dev = [
    "black==23.12.1",
    "pre-commit==3.6.0",
    "ruff==0.1.13",
    "pytest-aiohttp==1.0.5",
    "pytest-env==1.1.3",
    "pytest-mock==3.12.0",
    "faker==22.5.0",
]
prod = ["gunicorn==21.2.0"]

[tool.isort]
profile = "black"
